reuse `_safe_matmul` (#1014)
* additional `_safe_matmul`
* chlog
