+++ torchmetrics/functional/pairwise/linear.py
@@ -17,6 +17,7 @@ from torch import Tensor
 from typing_extensions import Literal
 
 from torchmetrics.functional.pairwise.helpers import _check_input, _reduce_distance_matrix
+from torchmetrics.utilities.compute import _safe_matmul
 
 
 def _pairwise_linear_similarity_update(
@@ -31,7 +32,7 @@ def _pairwise_linear_similarity_update(
     """
     x, y, zero_diagonal = _check_input(x, y, zero_diagonal)
 
-    distance = x @ y.T
+    distance = _safe_matmul(x, y)
     if zero_diagonal:
         distance.fill_diagonal_(0)
     return distance
